restructure and rename pages directory to app. Started readme.md
