--- .env.sample
@@ -1,4 +1,0 @@
-export POSTGRES_DB="pong"
-export POSTGRES_USER="pong"
-export POSTGRES_PASSWORD="pong"
-export POSTGRES_HOST="db"
