+++ .gitignore
@@ -19,3 +19,5 @@ venv.bak/
 
 # OSX
 .DS_Store
+
+db
