+++ front/README.md
@@ -1,0 +1,10 @@
+# Transcendence front
+
+## app directory
+Contains either layout directories or page directories. Where layout and page structure happens.
+
+## layout directory
+Each folder with name in parenthesis represents a layout, i.e. shared UI. Page folders nested within these folders share the same layout. Layout is defined in layout.js
+
+## page directory
+
