+++ front/app/(home)/View.js
